refactor: Dto 클래스 통합
