--- src/main/java/com/example/matdongsanserver/domain/story/dto/request/StoryUpdateRequest.java
@@ -1,17 +1,0 @@
-package com.example.matdongsanserver.domain.story.dto.request;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class StoryUpdateRequest {
-    private String title;
-    private Boolean isPublic;
-    private List<String> tags;
-}
