--- src/main/java/com/example/matdongsanserver/domain/story/dto/response/ChatGptResponse.java
@@ -1,16 +1,0 @@
-package com.example.matdongsanserver.domain.story.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
-
-import java.util.List;
-
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ChatGptResponse {
-    private String id;
-    private List<ChoiceResponse> choices;
-}
