--- src/main/java/com/example/matdongsanserver/domain/story/dto/response/ChoiceResponse.java
@@ -1,13 +1,0 @@
-package com.example.matdongsanserver.domain.story.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
-
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ChoiceResponse {
-    private MessageResponse message;
-}
